+++ packages/modeling/src/curves/bezier/lengths.d.ts
@@ -1,0 +1,3 @@
+import Bezier from './type'
+
+export function lengths(segments: number, bezier: Bezier): Array<number>