+++ packages/modeling/src/geometries/slice/toEdges.d.ts
@@ -1,3 +1,3 @@
 import { Slice } from './type'
 
-export function toEdges(slice: Slice): Slice['edges']
+export function toEdges(slice: Slice): Slice['contours']