+++ packages/modeling/src/geometries/slice/toVertices.d.ts
@@ -1,4 +1,4 @@
 import { Vec3 } from '../../maths/types'
 import { Slice } from './type'
 
-export function toPoints(slice: Slice): Array<Vec3>
+export function toVertices(slice: Slice): Array<Vec3>