+++ packages/modeling/src/operations/modifiers/index.d.ts
@@ -1,2 +1,3 @@
 export { generalize } from './generalize'
 export { snap } from './snap'
+export { retessellate } from './retessellate'